メールアドレスは不要

+++ flaskr/db/seeds/__init__.py
@@ -1,14 +1,19 @@
 import os
+from flaskr import app
 from flaskr.core import db
+from flaskr.models.user import User
+from flaskr.models.entry import Entry
 from flaskr.db.seeds.entry import entries
 from flaskr.db.seeds.user import users
 
 def init_db_with_seeds():
-    os.remove("flaskr/flaskr.db")
-    db.create_all()
+    if os.path.exists("flaskr/flaskr.db"):
+        os.remove("flaskr/flaskr.db")
 
-    db.session.bulk_save_objects(entries)
-    db.session.bulk_save_objects(users)
-    db.session.commit()
+    with app.app_context():
+        db.create_all()
+        db.session.bulk_save_objects([Entry(title=d[0], text=d[1]) for d in entries])
+        db.session.bulk_save_objects([User(name=d[0], password=d[1]) for d in users])
+        db.session.commit()
 
     print("データベースを初期化しました。")
